Changes in this commit: - The application now add fixed sessions of study - Decoupled the AndroidCalendar Class - Bug fixes
