--- app/src/main/java/com/gcw_rome_2014/saveme/calendar/InstanceProvider.java
@@ -1,7 +1,0 @@
-package com.gcw_rome_2014.saveme.calendar;
-
-/**
- * Created by Luigi on 03/01/2015.
- */
-public interface InstanceProvider {
-}
