--- app/src/main/java/com/gcw_rome_2014/saveme/calendar/android/AndroidInstance.java
@@ -1,10 +1,0 @@
-package com.gcw_rome_2014.saveme.calendar.android;
-
-import com.gcw_rome_2014.saveme.calendar.InstanceProvider;
-
-/**
- * Created by Luigi on 03/01/2015.
- */
-public class AndroidInstance implements InstanceProvider {
-
-}
