+++ app/src/main/java/com/gcw_rome_2014/saveme/calendar/provider/CalendarManager.java
@@ -1,0 +1,12 @@
+package com.gcw_rome_2014.saveme.calendar.provider;
+
+import java.util.Calendar;
+
+/**
+ * Created by Luigi Mortaro on 19/12/2014.
+ *
+ * This Interface has the scope to decouple the technology used to
+ */
+public interface CalendarManager {
+
+}
