+++ app/src/main/java/com/gcw_rome_2014/saveme/calendar/provider/EventManager.java
@@ -1,0 +1,14 @@
+package com.gcw_rome_2014.saveme.calendar.provider;
+
+import android.content.ContentValues;
+import android.net.Uri;
+import android.provider.CalendarContract;
+
+import java.util.Calendar;
+
+/**
+ * Created by Luigi on 03/01/2015.
+ */
+public interface EventManager {
+    public long addEvent(String eventName, String eventDescription, Calendar dateAndTime, int HoursPerDay);
+}
