+++ app/src/main/java/com/gcw_rome_2014/saveme/calendar/provider/InstanceManager.java
@@ -1,0 +1,7 @@
+package com.gcw_rome_2014.saveme.calendar.provider;
+
+/**
+ * Created by Luigi on 03/01/2015.
+ */
+public interface InstanceManager {
+}
